+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * CPU related facts.
 * `puppetresources` now exits with the proper error code.
 * `puppetresources` can now display some statistics about compilation times.
+* Bumped the version of the http-conduit dependency.
 ## Bugs fixed
 * Fixed dependencies so that builds with GHC 7.8-rc1 work.